+++ Source/Model/Game.py
@@ -1,11 +1,12 @@
 from concurrent.futures import thread
+from threading import Thread
 from Subject import Subject
 from Model.Colisiones import Colisiones
 from Model.GameObject import GameObject
 from Model.Tuberia import Tuberia
 import pygame
 import random
-from threading import Thread
+from PyQt5.QtCore import pyqtSignal
 
 
 class Game(Subject):
@@ -17,9 +18,11 @@ class Game(Subject):
     __ancho: int
     __alto: int
 
+    __gameoverSignal: pyqtSignal = pyqtSignal()
     __gameover: bool = False
 
     def __init__(self, ancho: int, alto: int):
+        super().__init__()
         pygame.init()
 
         self.__ancho = ancho
@@ -31,7 +34,10 @@ class Game(Subject):
         """
         Lanzar el game loop en un nuevo hilo
         """
+
+        #Comunicacion entre hilos ya que no puedo crear nueva view(GameoverView) desde el hilo nuevo
         gameLoopThread = Thread(target=self.__gameLoop)
+        self.__gameoverSignal.connect(self.onGameover)
         gameLoopThread.start()
 
     def getGameObjectsState(self) -> list[tuple[pygame.Surface, tuple[int, int]]]:
@@ -44,10 +50,16 @@ class Game(Subject):
         states.extend((tub.getSprite(), tub.getPosicion())
                       for tuple in self.__tuberias for tub in tuple)
 
-        return states    
+        return states
 
-    def isGameOver(self)-> bool:
+    def isGameOver(self) -> bool:
         return self.__gameover
+
+    def __initPajaro(self):
+        """
+        Inicializar pajaro en la posicion inicial
+        """
+        self.__pajaro.mover(self.__ancho / 5, self.__alto / 2)
 
     def __initTuberias(self):
         """
@@ -86,12 +98,6 @@ class Game(Subject):
 
         self.__tuberias.append((tuberiaInferior, tuberiaSuperior))
 
-    def __initPajaro(self):
-        """
-        Inicializar pajaro en la posicion inicial
-        """
-        self.__pajaro.mover(self.__ancho / 5, self.__alto / 2)
-
     def __gameLoop(self):
         """
         Game loop
@@ -106,9 +112,7 @@ class Game(Subject):
                     tuberia.actualizar(deltaTime)
 
                     if(Colisiones.colisiona(tuberia, self.__pajaro)):
-                        self.__gameover = True
-
-                        self._notify(self)
+                        self.__gameoverSignal.emit()
                         return
 
                 if(Colisiones.parTuberiasAfuera(parTuberias)):
@@ -120,3 +124,7 @@ class Game(Subject):
             self._notify(self)
 
             self.__relojFrames.tick(60)
+
+    def onGameover(self):
+        self.__gameover = True
+        self._notify(self)
