+++ Source/Views/GameoverView.py
@@ -1,10 +1,8 @@
-from Source.Subject import Subject
-from View import View
+from Views.View import View
 from Controllers.Controller import Controller
-from PyQt5.QtWidgets import QMainWindow
 
 
 class GameoverView(View):
 
     def __init__(self,controller: Controller):
-        super().__init__(self,controller)
+        super().__init__(controller)
